backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
supabase==2.28.3
postgrest==2.28.3
python-dotenv==1.2.2
pydantic==2.12.5
pydantic-settings==2.13.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
httpx==0.28.1
redis==7.3.0
openai==2.29.0
boto3==1.42.74
moviepy==2.2.1
Pillow==11.3.0
numpy==2.4.3
celery==5.6.2
websockets==15.0
sendgrid==6.12.5
twilio==9.10.3
stripe==14.4.1
python-dateutil==2.9.0.post0
alembic==1.12.1
psycopg2-binary==2.9.9
reportlab==4.4.10
pandas==3.0.1
email-validator==2.1.1